+++ Sample/represent-sample/route-sample.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
