+++ Sample/represent-sample/route-sample.xcworkspace/contents.xcworkspacedata
@@ -5,6 +5,6 @@
       location = "group:../..">
    </FileRef>
    <FileRef
-      location = "group:represent-sample.xcodeproj">
+      location = "group:/Users/jsilver/Developer/project/repository/Route/Sample/represent-sample/route-sample.xcodeproj">
    </FileRef>
 </Workspace>
