+++ Sample/represent-sample/route-sample.xcworkspace/xcshareddata/swiftpm/Package.resolved
