+++ Sample/represent-sample/route-sample/Resource/Assets.xcassets/AppIcon.appiconset/Contents.json
