+++ Sample/represent-sample/route-sample/Source/Main/View/Cell/MainActionTableViewCell.swift
@@ -1,30 +1,20 @@
 //
-//  MainTableViewCell.swift
-//  represent-sample
+//  MainActionTableViewCell.swift
+//  route-sample
 //
-//  Created by JSilver on 2022/11/11.
+//  Created by JSilver on 2022/11/12.
 //  Copyright © 2022 JSilver. All rights reserved.
 //
 
 import UIKit
 import SnapKit
 
-class MainTableViewCell: UITableViewCell {
+class MainActionTableViewCell: UITableViewCell {
     // MARK: - View
     private let titleLable: UILabel = {
         let view = UILabel()
         view.font = .preferredFont(forTextStyle: .body)
-        view.setContentCompressionResistancePriority(.required, for: .horizontal)
-        
-        return view
-    }()
-    
-    private let descriptionLabel: UILabel = {
-        let view = UILabel()
-        view.font = .preferredFont(forTextStyle: .body)
-        view.textColor = .lightGray
-        view.textAlignment = .right
-        view.numberOfLines = 0
+        view.textColor = view.tintColor
         
         return view
     }()
@@ -46,13 +36,11 @@ class MainTableViewCell: UITableViewCell {
         super.prepareForReuse()
         
         titleLable.text = nil
-        descriptionLabel.text = nil
     }
     
     // MARK: - Public
-    public func configure(title: String, description: String) {
+    public func configure(title: String) {
         titleLable.text = title
-        descriptionLabel.text = description
     }
     
     // MARK: - Private
@@ -65,28 +53,16 @@ class MainTableViewCell: UITableViewCell {
     private func setUpLayout() {
         [
             titleLable,
-            descriptionLabel
         ]
             .forEach { contentView.addSubview($0) }
         
         titleLable.snp.makeConstraints {
-            $0.top.equalToSuperview()
+            $0.top.bottom
+                .equalToSuperview()
                 .inset(10)
-            $0.bottom.lessThanOrEqualToSuperview()
-                .inset(10)
-            $0.leading.equalToSuperview()
-                .inset(10)
-        }
-        
-        descriptionLabel.snp.makeConstraints {
-            $0.top.equalToSuperview()
-                .inset(10)
-            $0.trailing.equalToSuperview()
-                .inset(10)
-            $0.bottom.lessThanOrEqualToSuperview()
-                .inset(10)
-            $0.leading.equalTo(titleLable.snp.trailing)
-                .offset(16)
+            $0.leading.trailing
+                .equalToSuperview()
+                .inset(16)
         }
     }
