--- Tests/RepresentTests/RepresentTests.swift
@@ -1,13 +1,0 @@
-//
-//  UIViewController+Represent.swift
-//  
-//
-//  Created by jsilver on 2020/06/19.
-//
-
-import XCTest
-@testable import Represent
-
-final class RepresentTests: XCTestCase {
-    
-}
