+++ Tests/RouteTests/RouteTests.swift
@@ -1,0 +1,13 @@
+//
+//  RouteTests.swift
+//  
+//
+//  Created by jsilver on 2020/06/19.
+//
+
+import XCTest
+@testable import Route
+
+final class RouteTests: XCTestCase {
+    
+}
